Rebuild the runtime in order to support interop
The intention was to build similar interop with Python
as it is in Clojure with Java. This way would be possible
to access builtin implementation of stuff without to jack
things into the application scope.

Abstracted some stuff to a module `core.lisp`, which is not
exactly a module yet, since is not isolated (it can't be) from
the application scope.

Need to add macros and consequently quoting in order to support
more advanced Lisp features. I would probably need to implement
some type of "meta" feature on the declare function (if I can even
call it that way) so the runtime understands it as a macro.

NOTE: Need refactoring.